+++ app/accesData/entities/User.scala
@@ -1,3 +1,3 @@
 package accesData.entities
 
-case class User(id: Long, name: String, email: String, balance: Int)
+case class User(id: Long, name: String, email: String, balance: BigDecimal)
